+++ inc/styleguide.php
@@ -50,6 +50,7 @@ array_walk($telabotanica_modules, function ($module) {
  * Liste de tous les composants accessibles depuis les fonctions du styleguide
  */
 $telabotanica_blocks = [
+  'focus',
   'list-features',
   'main-features',
 ];